Migrate series plugin and companions into series component
