+++ flexget/api/__init__.py
@@ -3,6 +3,6 @@ from __future__ import unicode_literals, division, absolute_import  # noqa
 from .app import api_app, api, APIResource, APIClient  # noqa
 from .core import authentication, cached, database, plugins, server, tasks, user, format_checker  # noqa
 from .plugins import (
-    series, trakt_lookup, tvdb_lookup, tvmaze_lookup, tmdb_lookup, irc, status, pending_list, pending, entry_list,
+    trakt_lookup, tvdb_lookup, tvmaze_lookup, tmdb_lookup, irc, status, pending_list, pending, entry_list,
     failed, imdb_lookup, movie_list, rejected, schedule, variables  # noqa
 )
