+++ flexget/components/series/__init__.py
@@ -1,0 +1,12 @@
+# Import plugins when component is imported
+from . import api
+from . import cli
+from . import series
+from . import configure_series
+from . import next_series_episodes
+from . import next_series_seasons
+from . import all_series
+from . import metainfo_series
+from . import series_premiere
+from . import series_forget
+from . import internal_estimator
