+++ flexget/components/series/api.py
@@ -1,7 +1,7 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import copy
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from math import ceil
 
 from flask import jsonify
@@ -9,15 +9,35 @@ from flask import request
 from flask_restplus import inputs
 from sqlalchemy.orm.exc import NoResultFound
 
+from flexget import plugin
 from flexget.api import api, APIClient, APIResource
-from flexget.api.app import NotFoundError, Conflict, BadRequest, base_message_schema, success_response, etag, \
-    pagination_headers
+from flexget.api.app import (
+    NotFoundError,
+    Conflict,
+    BadRequest,
+    base_message_schema,
+    success_response,
+    etag,
+    pagination_headers,
+)
 from flexget.event import fire_event
 from flexget.plugin import PluginError
-from flexget.plugins.filter import series
+from . import db
+from .utils import normalize_series_name
 
-from flexget.api.plugins.tvmaze_lookup import ObjectsContainer as tvmaze
-from flexget.api.plugins.tvdb_lookup import ObjectsContainer as tvdb
+try:
+    # NOTE: Importing other plugins is discouraged!
+    from flexget.api.plugins.tvdb_lookup import ObjectsContainer as tvdb
+except ImportError:
+    raise plugin.DependencyError(issued_by=__name__, missing='tvdb_lookup')
+
+from flexget import plugin
+
+try:
+    # NOTE: Importing other plugins is discouraged!
+    from flexget.api.plugins.tvmaze_lookup import ObjectsContainer as tvmaze
+except ImportError:
+    raise plugin.DependencyError(issued_by=__name__, missing='tvmaze_lookup')
 
 series_api = api.namespace('series', description='Flexget Series operations')
 
@@ -27,12 +47,12 @@ def series_details(show, begin=False, latest=False):
         'id': show.id,
         'name': show.name,
         'alternate_names': [n.alt_name for n in show.alternate_names],
-        'in_tasks': [_show.name for _show in show.in_tasks]
+        'in_tasks': [_show.name for _show in show.in_tasks],
     }
     if begin:
         series_dict['begin_episode'] = show.begin.to_dict() if show.begin else None
     if latest:
-        latest_entity = series.get_latest_release(show)
+        latest_entity = db.get_latest_release(show)
         series_dict['latest_entity'] = latest_entity.to_dict() if latest_entity else None
         if latest_entity:
             series_dict['latest_entity']['latest_release'] = latest_entity.latest_release.to_dict()
@@ -49,9 +69,17 @@ class ObjectsContainer(object):
             'quality': {'type': 'string'},
             'proper_count': {'type': 'integer'},
             'first_seen': {'type': 'string', 'format': 'date-time'},
-            'episode_id': {'type': 'integer'}
+            'episode_id': {'type': 'integer'},
         },
-        'required': ['id', 'title', 'downloaded', 'quality', 'proper_count', 'first_seen', 'episode_id']
+        'required': [
+            'id',
+            'title',
+            'downloaded',
+            'quality',
+            'proper_count',
+            'first_seen',
+            'episode_id',
+        ],
     }
 
     season_release_object = copy.deepcopy(episode_release_object)
@@ -75,10 +103,19 @@ class ObjectsContainer(object):
             "season": {'type': 'integer'},
             "series_id": {'type': 'integer'},
             "number_of_releases": {'type': 'integer'},
-            'latest_release': episode_release_object
+            'latest_release': episode_release_object,
         },
-        'required': ['first_seen', 'id', 'identified_by', 'identifier', 'premiere', 'number', 'season',
-                     'series_id', 'number_of_releases']
+        'required': [
+            'first_seen',
+            'id',
+            'identified_by',
+            'identifier',
+            'premiere',
+            'number',
+            'season',
+            'series_id',
+            'number_of_releases',
+        ],
     }
 
     season_object = copy.deepcopy(episode_object)
@@ -97,13 +134,13 @@ class ObjectsContainer(object):
                 'type': 'object',
                 'properties': {
                     'tvmaze': tvmaze.tvmaze_series_object,
-                    'tvdb': tvdb.tvdb_series_object
-                }
+                    'tvdb': tvdb.tvdb_series_object,
+                },
             },
             'latest_episode': episode_object,
-            'begin_episode': episode_object
+            'begin_episode': episode_object,
         },
-        'required': ['id', 'name', 'alternate_names', 'in_tasks']
+        'required': ['id', 'name', 'alternate_names', 'in_tasks'],
     }
 
     series_list_schema = {'type': 'array', 'items': single_series_object}
@@ -115,15 +152,11 @@ class ObjectsContainer(object):
     series_edit_object = {
         'type': 'object',
         'properties': {
-            'begin_episode':
-                {'type': ['string', 'integer'], 'format': 'episode_or_season_id'},
-            'alternate_names': {'type': 'array', 'items': {'type': 'string'}}
+            'begin_episode': {'type': ['string', 'integer'], 'format': 'episode_or_season_id'},
+            'alternate_names': {'type': 'array', 'items': {'type': 'string'}},
         },
-        'anyOf': [
-            {'required': ['begin_episode']},
-            {'required': ['alternate_names']}
-        ],
-        'additionalProperties:': False
+        'anyOf': [{'required': ['begin_episode']}, {'required': ['alternate_names']}],
+        'additionalProperties:': False,
     }
 
     series_input_object = copy.deepcopy(series_edit_object)
@@ -143,28 +176,51 @@ episode_schema = api.schema_model('episode_item', ObjectsContainer.episode_objec
 season_list_schema = api.schema_model('season_list', ObjectsContainer.seasons_list_schema)
 season_schema = api.schema_model('episode_item', ObjectsContainer.season_object)
 
-episode_release_schema = api.schema_model('release_schema', ObjectsContainer.episode_release_object)
-episode_release_list_schema = api.schema_model('release_list_schema', ObjectsContainer.episode_release_list_schema)
+episode_release_schema = api.schema_model(
+    'release_schema', ObjectsContainer.episode_release_object
+)
+episode_release_list_schema = api.schema_model(
+    'release_list_schema', ObjectsContainer.episode_release_list_schema
+)
 
 season_release_schema = api.schema_model('release_schema', ObjectsContainer.season_release_object)
-season_release_list_schema = api.schema_model('release_list_schema', ObjectsContainer.season_release_list_schema)
+season_release_list_schema = api.schema_model(
+    'release_list_schema', ObjectsContainer.season_release_list_schema
+)
 
 base_series_parser = api.parser()
-base_series_parser.add_argument('begin', type=inputs.boolean, default=True, help='Show series begin episode')
-base_series_parser.add_argument('latest', type=inputs.boolean, default=True,
-                                help='Show series latest downloaded episode and release')
+base_series_parser.add_argument(
+    'begin', type=inputs.boolean, default=True, help='Show series begin episode'
+)
+base_series_parser.add_argument(
+    'latest',
+    type=inputs.boolean,
+    default=True,
+    help='Show series latest downloaded episode and release',
+)
 
 sort_choices = ('show_name', 'last_download_date')
 series_list_parser = api.pagination_parser(base_series_parser, sort_choices=sort_choices)
-series_list_parser.add_argument('in_config', choices=('configured', 'unconfigured', 'all'), default='configured',
-                                help="Filter list if shows are currently in configuration.")
-series_list_parser.add_argument('premieres', type=inputs.boolean, default=False,
-                                help="Filter by downloaded premieres only.")
-series_list_parser.add_argument('lookup', choices=('tvdb', 'tvmaze'), action='append',
-                                help="Get lookup result for every show by sending another request to lookup API")
+series_list_parser.add_argument(
+    'in_config',
+    choices=('configured', 'unconfigured', 'all'),
+    default='configured',
+    help="Filter list if shows are currently in configuration.",
+)
+series_list_parser.add_argument(
+    'premieres', type=inputs.boolean, default=False, help="Filter by downloaded premieres only."
+)
+series_list_parser.add_argument(
+    'lookup',
+    choices=('tvdb', 'tvmaze'),
+    action='append',
+    help="Get lookup result for every show by sending another request to lookup API",
+)
 series_list_parser.add_argument('query', help="Search by name based on the query")
 
-ep_identifier_doc = "'episode_identifier' should be one of SxxExx, integer or date formatted such as 2012-12-12"
+ep_identifier_doc = (
+    "'episode_identifier' should be one of SxxExx, integer or date formatted such as 2012-12-12"
+)
 
 
 @series_api.route('/')
@@ -186,7 +242,7 @@ class SeriesAPI(APIResource):
         per_page = args['per_page']
         sort_by = args['sort_by']
         sort_order = args['order']
-        name = series.normalize_series_name(args['query']) if args['query'] else None
+        name = normalize_series_name(args['query']) if args['query'] else None
 
         # Handle max size limit
         if per_page > 100:
@@ -210,16 +266,16 @@ class SeriesAPI(APIResource):
             'sort_by': sort_by,
             'descending': descending,
             'session': session,
-            'name': name
+            'name': name,
         }
 
-        total_items = series.get_series_summary(count=True, **kwargs)
+        total_items = db.get_series_summary(count=True, **kwargs)
 
         if not total_items:
             return jsonify([])
 
         series_list = []
-        for s in series.get_series_summary(**kwargs):
+        for s in db.get_series_summary(**kwargs):
             series_object = series_details(s, begin, latest)
             series_list.append(series_object)
 
@@ -262,21 +318,21 @@ class SeriesAPI(APIResource):
         data = request.json
         series_name = data.get('name')
 
-        normalized_name = series.normalize_series_name(series_name)
-        matches = series.shows_by_exact_name(normalized_name, session=session)
+        normalized_name = normalize_series_name(series_name)
+        matches = db.shows_by_exact_name(normalized_name, session=session)
         if matches:
             raise Conflict('Show `%s` already exist in DB' % series_name)
-        show = series.Series()
+        show = db.Series()
         show.name = series_name
         session.add(show)
 
         ep_id = data.get('begin_episode')
         alt_names = data.get('alternate_names')
         if ep_id:
-            series.set_series_begin(show, ep_id)
+            db.set_series_begin(show, ep_id)
         if alt_names:
             try:
-                series.set_alt_names(alt_names, show, session)
+                db.set_alt_names(alt_names, show, session)
             except PluginError as e:
                 # Alternate name already exist for a different show
                 raise Conflict(e.value)
@@ -287,15 +343,17 @@ class SeriesAPI(APIResource):
 
 
 @series_api.route('/search/<string:name>/')
-@api.doc(description='Searches for a show in the DB via its name. Returns a list of matching shows.')
+@api.doc(
+    description='Searches for a show in the DB via its name. Returns a list of matching shows.'
+)
 class SeriesGetShowsAPI(APIResource):
     @etag
     @api.response(200, 'Show list retrieved successfully', series_list_schema)
     @api.doc(params={'name': 'Name of the show(s) to search'}, parser=base_series_parser)
     def get(self, name, session):
         """ List of shows matching lookup name """
-        name = series.normalize_series_name(name)
-        matches = series.shows_by_name(name, session=session)
+        name = normalize_series_name(name)
+        matches = db.shows_by_name(name, session=session)
 
         args = series_list_parser.parse_args()
         begin = args.get('begin')
@@ -309,9 +367,13 @@ class SeriesGetShowsAPI(APIResource):
 
 
 delete_parser = api.parser()
-delete_parser.add_argument('forget', type=inputs.boolean, default=False,
-                           help="Enabling this will fire a 'forget' event that will delete the downloaded releases "
-                                "from the entire DB, enabling to re-download them")
+delete_parser.add_argument(
+    'forget',
+    type=inputs.boolean,
+    default=False,
+    help="Enabling this will fire a 'forget' event that will delete the downloaded releases "
+    "from the entire DB, enabling to re-download them",
+)
 
 
 @series_api.route('/<int:show_id>/')
@@ -324,7 +386,7 @@ class SeriesShowAPI(APIResource):
     def get(self, show_id, session):
         """ Get show details by ID """
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('Show with ID %s not found' % show_id)
 
@@ -335,40 +397,43 @@ class SeriesShowAPI(APIResource):
         return jsonify(series_details(show, begin, latest))
 
     @api.response(200, 'Removed series from DB', model=base_message_schema)
-    @api.doc(description='Delete a specific show using its ID',
-             parser=delete_parser)
+    @api.doc(description='Delete a specific show using its ID', parser=delete_parser)
     def delete(self, show_id, session):
         """ Remove series from DB """
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('Show with ID %s not found' % show_id)
 
         name = show.name
         args = delete_parser.parse_args()
-        series.remove_series(name, forget=args.get('forget'))
+        db.remove_series(name, forget=args.get('forget'))
 
         return success_response('successfully removed series %s from DB' % show_id)
 
-    @api.response(200, 'Episodes for series will be accepted starting with ep_id', show_details_schema)
+    @api.response(
+        200, 'Episodes for series will be accepted starting with ep_id', show_details_schema
+    )
     @api.response(Conflict)
     @api.validate(series_edit_schema, description=ep_identifier_doc)
-    @api.doc(description='Set a begin episode or alternate names using a show ID. Note that alternate names override '
-                         'the existing names (if name does not belong to a different show).')
+    @api.doc(
+        description='Set a begin episode or alternate names using a show ID. Note that alternate names override '
+        'the existing names (if name does not belong to a different show).'
+    )
     def put(self, show_id, session):
         """ Set the initial episode of an existing show """
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('Show with ID %s not found' % show_id)
         data = request.json
         ep_id = data.get('begin_episode')
         alt_names = data.get('alternate_names')
         if ep_id:
-            series.set_series_begin(show, ep_id)
+            db.set_series_begin(show, ep_id)
         if alt_names:
             try:
-                series.set_alt_names(alt_names, show, session)
+                db.set_alt_names(alt_names, show, session)
             except PluginError as e:
                 # Alternate name already exist for a different show
                 raise Conflict(e.value)
@@ -380,8 +445,10 @@ entity_parser = api.pagination_parser(add_sort=True)
 
 @api.response(NotFoundError)
 @series_api.route('/<int:show_id>/seasons/')
-@api.doc(params={'show_id': 'ID of the show'},
-         description='The \'Series-ID\' header will be appended to the result headers')
+@api.doc(
+    params={'show_id': 'ID of the show'},
+    description='The \'Series-ID\' header will be appended to the result headers',
+)
 class SeriesSeasonsAPI(APIResource):
     @etag
     @api.response(200, 'Seasons retrieved successfully for show', season_list_schema)
@@ -404,24 +471,19 @@ class SeriesSeasonsAPI(APIResource):
         start = per_page * (page - 1)
         stop = start + per_page
 
-        kwargs = {
-            'start': start,
-            'stop': stop,
-            'descending': descending,
-            'session': session
-        }
+        kwargs = {'start': start, 'stop': stop, 'descending': descending, 'session': session}
 
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
 
-        total_items = series.show_seasons(show, count=True, session=session)
+        total_items = db.show_seasons(show, count=True, session=session)
 
         if not total_items:
             return jsonify([])
 
-        seasons = [season.to_dict() for season in series.show_seasons(show, **kwargs)]
+        seasons = [season.to_dict() for season in db.show_seasons(show, **kwargs)]
 
         total_pages = int(ceil(total_items / float(per_page)))
 
@@ -445,18 +507,20 @@ class SeriesSeasonsAPI(APIResource):
         return rsp
 
     @api.response(200, 'Successfully forgotten all seasons from show', model=base_message_schema)
-    @api.doc(description='Delete all show seasons via its ID. Deleting a season will mark it as wanted again',
-             parser=delete_parser)
+    @api.doc(
+        description='Delete all show seasons via its ID. Deleting a season will mark it as wanted again',
+        parser=delete_parser,
+    )
     def delete(self, show_id, session):
         """ Deletes all seasons of a show"""
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         args = delete_parser.parse_args()
         forget = args.get('forget')
         for season in show.seasons:
-            series.remove_series_entity(show.name, season.identifier, forget)
+            db.remove_series_entity(show.name, season.identifier, forget)
         return success_response('successfully removed all series %s seasons from DB' % show_id)
 
 
@@ -471,14 +535,14 @@ class SeriesSeasonsAPI(APIResource):
     def get(self, show_id, season_id, session):
         """ Get season by show ID and season ID"""
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            season = series.season_by_id(season_id, session)
+            season = db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
 
         rsp = jsonify(season.to_dict())
@@ -488,32 +552,38 @@ class SeriesSeasonsAPI(APIResource):
         return rsp
 
     @api.response(200, 'Season successfully forgotten for show', model=base_message_schema)
-    @api.doc(description='Delete a specific season via its ID and show ID. Deleting a season will mark it as '
-                         'wanted again',
-             parser=delete_parser)
+    @api.doc(
+        description='Delete a specific season via its ID and show ID. Deleting a season will mark it as '
+        'wanted again',
+        parser=delete_parser,
+    )
     def delete(self, show_id, season_id, session):
         """ Forgets season by show ID and season ID """
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            season = series.season_by_id(season_id, session)
+            season = db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
 
         args = delete_parser.parse_args()
-        series.remove_series_entity(show.name, season.identifier, args.get('forget'))
+        db.remove_series_entity(show.name, season.identifier, args.get('forget'))
 
-        return success_response('successfully removed season %s from show %s' % (season_id, show_id))
+        return success_response(
+            'successfully removed season %s from show %s' % (season_id, show_id)
+        )
 
 
 @api.response(NotFoundError)
 @series_api.route('/<int:show_id>/episodes/')
-@api.doc(params={'show_id': 'ID of the show'},
-         description='The \'Series-ID\' header will be appended to the result headers')
+@api.doc(
+    params={'show_id': 'ID of the show'},
+    description='The \'Series-ID\' header will be appended to the result headers',
+)
 class SeriesEpisodesAPI(APIResource):
     @etag
     @api.response(200, 'Episodes retrieved successfully for show', episode_list_schema)
@@ -536,24 +606,19 @@ class SeriesEpisodesAPI(APIResource):
         start = per_page * (page - 1)
         stop = start + per_page
 
-        kwargs = {
-            'start': start,
-            'stop': stop,
-            'descending': descending,
-            'session': session
-        }
+        kwargs = {'start': start, 'stop': stop, 'descending': descending, 'session': session}
 
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
 
-        total_items = series.show_episodes(show, count=True, session=session)
+        total_items = db.show_episodes(show, count=True, session=session)
 
         if not total_items:
             return jsonify([])
 
-        episodes = [episode.to_dict() for episode in series.show_episodes(show, **kwargs)]
+        episodes = [episode.to_dict() for episode in db.show_episodes(show, **kwargs)]
 
         total_pages = int(ceil(total_items / float(per_page)))
 
@@ -577,18 +642,20 @@ class SeriesEpisodesAPI(APIResource):
         return rsp
 
     @api.response(200, 'Successfully forgotten all episodes from show', model=base_message_schema)
-    @api.doc(description='Delete all show episodes via its ID. Deleting an episode will mark it as wanted again',
-             parser=delete_parser)
+    @api.doc(
+        description='Delete all show episodes via its ID. Deleting an episode will mark it as wanted again',
+        parser=delete_parser,
+    )
     def delete(self, show_id, session):
         """ Deletes all episodes of a show"""
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         args = delete_parser.parse_args()
         forget = args.get('forget')
         for episode in show.episodes:
-            series.remove_series_entity(show.name, episode.identifier, forget)
+            db.remove_series_entity(show.name, episode.identifier, forget)
         return success_response('successfully removed all series %s episodes from DB' % show_id)
 
 
@@ -603,14 +670,14 @@ class SeriesEpisodeAPI(APIResource):
     def get(self, show_id, ep_id, session):
         """ Get episode by show ID and episode ID"""
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            episode = series.episode_by_id(ep_id, session)
+            episode = db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
 
         rsp = jsonify(episode.to_dict())
@@ -620,64 +687,78 @@ class SeriesEpisodeAPI(APIResource):
         return rsp
 
     @api.response(200, 'Episode successfully forgotten for show', model=base_message_schema)
-    @api.doc(description='Delete a specific episode via its ID and show ID. Deleting an episode will mark it as '
-                         'wanted again',
-             parser=delete_parser)
+    @api.doc(
+        description='Delete a specific episode via its ID and show ID. Deleting an episode will mark it as '
+        'wanted again',
+        parser=delete_parser,
+    )
     def delete(self, show_id, ep_id, session):
         """ Forgets episode by show ID and episode ID """
         try:
-            show = series.show_by_id(show_id, session=session)
+            show = db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            episode = series.episode_by_id(ep_id, session)
+            episode = db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
 
         args = delete_parser.parse_args()
-        series.remove_series_entity(show.name, episode.identifier, args.get('forget'))
+        db.remove_series_entity(show.name, episode.identifier, args.get('forget'))
 
         return success_response('successfully removed episode %s from show %s' % (ep_id, show_id))
 
 
 release_base_parser = api.parser()
-release_base_parser.add_argument('downloaded', type=inputs.boolean, help='Filter between release status')
+release_base_parser.add_argument(
+    'downloaded', type=inputs.boolean, help='Filter between release status'
+)
 
 sort_choices = ('first_seen', 'downloaded', 'proper_count', 'title')
 release_list_parser = api.pagination_parser(release_base_parser, sort_choices)
 
 release_delete_parser = release_base_parser.copy()
-release_delete_parser.add_argument('forget', type=inputs.boolean, default=False,
-                                   help="Enabling this will for 'forget' event that will delete the downloaded"
-                                        " releases from the entire DB, enabling to re-download them")
+release_delete_parser.add_argument(
+    'forget',
+    type=inputs.boolean,
+    default=False,
+    help="Enabling this will for 'forget' event that will delete the downloaded"
+    " releases from the entire DB, enabling to re-download them",
+)
 
 
 @api.response(NotFoundError)
 @api.response(BadRequest)
 @series_api.route('/<int:show_id>/seasons/<int:season_id>/releases/')
-@api.doc(params={'show_id': 'ID of the show', 'season_id': 'Seasons ID'},
-         description='Releases are any seen entries that match the seasons. \n'
-                     'The \'Series-ID\' header will be appended to the result headers.\n'
-                     'The \'Season-ID\' header will be appended to the result headers.')
+@api.doc(
+    params={'show_id': 'ID of the show', 'season_id': 'Seasons ID'},
+    description='Releases are any seen entries that match the seasons. \n'
+    'The \'Series-ID\' header will be appended to the result headers.\n'
+    'The \'Season-ID\' header will be appended to the result headers.',
+)
 class SeriesSeasonsReleasesAPI(APIResource):
     @etag
     @api.response(200, 'Releases retrieved successfully for season', season_release_list_schema)
-    @api.doc(description='Get all matching releases for a specific season of a specific show.',
-             parser=release_list_parser)
+    @api.doc(
+        description='Get all matching releases for a specific season of a specific show.',
+        parser=release_list_parser,
+    )
     def get(self, show_id, season_id, session):
         """ Get all season releases by show ID and season ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            season = series.season_by_id(season_id, session)
+            season = db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
-        if not series.season_in_show(show_id, season_id):
-            raise BadRequest('seasons with id %s does not belong to show %s' % (season_id, show_id))
+        if not db.season_in_show(show_id, season_id):
+            raise BadRequest(
+                'seasons with id %s does not belong to show %s' % (season_id, show_id)
+            )
 
         args = release_list_parser.parse_args()
         # Filter params
@@ -705,14 +786,14 @@ class SeriesSeasonsReleasesAPI(APIResource):
             'sort_by': sort_by,
             'descending': descending,
             'season': season,
-            'session': session
+            'session': session,
         }
 
         # Total number of releases
-        total_items = series.get_season_releases(count=True, **kwargs)
+        total_items = db.get_season_releases(count=True, **kwargs)
 
         # Release items
-        release_items = [release.to_dict() for release in series.get_season_releases(**kwargs)]
+        release_items = [release.to_dict() for release in db.get_season_releases(**kwargs)]
 
         # Total number of pages
         total_pages = int(ceil(total_items / float(per_page)))
@@ -738,143 +819,165 @@ class SeriesSeasonsReleasesAPI(APIResource):
         return rsp
 
     @api.response(200, 'Successfully deleted all releases for season', model=base_message_schema)
-    @api.doc(description='Delete all releases for a specific season of a specific show.',
-             parser=release_delete_parser)
+    @api.doc(
+        description='Delete all releases for a specific season of a specific show.',
+        parser=release_delete_parser,
+    )
     def delete(self, show_id, season_id, session):
         """ Deletes all season releases by show ID and season ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            season = series.season_by_id(season_id, session)
+            season = db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('seasons with ID %s not found' % season_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
 
         args = release_delete_parser.parse_args()
         downloaded = args.get('downloaded') is True if args.get('downloaded') is not None else None
         release_items = []
         for release in season.releases:
-            if downloaded and release.downloaded or downloaded is False and not release.downloaded or not downloaded:
+            if (
+                downloaded
+                and release.downloaded
+                or downloaded is False
+                and not release.downloaded
+                or not downloaded
+            ):
                 release_items.append(release)
 
         for release in release_items:
             if args.get('forget'):
                 fire_event('forget', release.title)
-            series.delete_season_release_by_id(release.id)
-        return success_response('successfully deleted all releases for season %s from show %s' % (season_id, show_id))
+            db.delete_season_release_by_id(release.id)
+        return success_response(
+            'successfully deleted all releases for season %s from show %s' % (season_id, show_id)
+        )
 
-    @api.response(200, 'Successfully reset all downloaded releases for season', model=base_message_schema)
-    @api.doc(description='Resets all of the downloaded releases of an season, clearing the quality to be downloaded '
-                         'again,')
+    @api.response(
+        200, 'Successfully reset all downloaded releases for season', model=base_message_schema
+    )
+    @api.doc(
+        description='Resets all of the downloaded releases of an season, clearing the quality to be downloaded '
+        'again,'
+    )
     def put(self, show_id, season_id, session):
         """ Marks all downloaded season releases as not downloaded """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            season = series.season_by_id(season_id, session)
+            season = db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
 
         for release in season.releases:
             if release.downloaded:
                 release.downloaded = False
 
-        return success_response('successfully reset download status for all releases for season %s from show %s'
-                                % (season_id, show_id))
+        return success_response(
+            'successfully reset download status for all releases for season %s from show %s'
+            % (season_id, show_id)
+        )
 
 
 @api.response(NotFoundError)
 @api.response(BadRequest)
 @series_api.route('/<int:show_id>/seasons/<int:season_id>/releases/<int:rel_id>/')
-@api.doc(params={'show_id': 'ID of the show', 'season_id': 'Season ID', 'rel_id': 'Release ID'},
-         description='Releases are any seen entries that match the season. \n'
-                     'The \'Series-ID\' header will be appended to the result headers.\n'
-                     'The \'Season-ID\' header will be appended to the result headers.'
-         )
+@api.doc(
+    params={'show_id': 'ID of the show', 'season_id': 'Season ID', 'rel_id': 'Release ID'},
+    description='Releases are any seen entries that match the season. \n'
+    'The \'Series-ID\' header will be appended to the result headers.\n'
+    'The \'Season-ID\' header will be appended to the result headers.',
+)
 class SeriesSeasonReleaseAPI(APIResource):
     @etag
     @api.response(200, 'Release retrieved successfully for season', season_release_schema)
-    @api.doc(description='Get a specific downloaded release for a specific season of a specific show')
+    @api.doc(
+        description='Get a specific downloaded release for a specific season of a specific show'
+    )
     def get(self, show_id, season_id, rel_id, session):
         """ Get season release by show ID, season ID and release ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            series.season_by_id(season_id, session)
+            db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
         try:
-            release = series.season_release_by_id(rel_id, session)
+            release = db.season_release_by_id(rel_id, session)
         except NoResultFound:
             raise NotFoundError('release with ID %s not found' % rel_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
-        if not series.release_in_season(season_id, rel_id):
+        if not db.release_in_season(season_id, rel_id):
             raise BadRequest('release id %s does not belong to season %s' % (rel_id, season_id))
 
         rsp = jsonify(release.to_dict())
-        rsp.headers.extend({
-            'Series-ID': show_id,
-            'Season-ID': season_id
-        })
+        rsp.headers.extend({'Series-ID': show_id, 'Season-ID': season_id})
         return rsp
 
     @api.response(200, 'Release successfully deleted', model=base_message_schema)
-    @api.doc(description='Delete a specific releases for a specific season of a specific show.',
-             parser=delete_parser)
+    @api.doc(
+        description='Delete a specific releases for a specific season of a specific show.',
+        parser=delete_parser,
+    )
     def delete(self, show_id, season_id, rel_id, session):
         """ Delete episode release by show ID, season ID and release ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            series.season_by_id(season_id, session)
+            db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
         try:
-            release = series.season_release_by_id(rel_id, session)
+            release = db.season_release_by_id(rel_id, session)
         except NoResultFound:
             raise NotFoundError('release with ID %s not found' % rel_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
-        if not series.release_in_season(season_id, rel_id):
+        if not db.release_in_season(season_id, rel_id):
             raise BadRequest('release id %s does not belong to season %s' % (rel_id, season_id))
         args = delete_parser.parse_args()
         if args.get('forget'):
             fire_event('forget', release.title)
 
-        series.delete_season_release_by_id(rel_id)
-        return success_response('successfully deleted release %d from season %d' % (rel_id, season_id))
+        db.delete_season_release_by_id(rel_id)
+        return success_response(
+            'successfully deleted release %d from season %d' % (rel_id, season_id)
+        )
 
     @api.response(200, 'Successfully reset downloaded release status', model=season_release_schema)
-    @api.doc(description='Resets the downloaded release status, clearing the quality to be downloaded again')
+    @api.doc(
+        description='Resets the downloaded release status, clearing the quality to be downloaded again'
+    )
     def put(self, show_id, season_id, rel_id, session):
         """ Resets a downloaded release status """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            series.season_by_id(season_id, session)
+            db.season_by_id(season_id, session)
         except NoResultFound:
             raise NotFoundError('season with ID %s not found' % season_id)
         try:
-            release = series.season_release_by_id(rel_id, session)
+            release = db.season_release_by_id(rel_id, session)
         except NoResultFound:
             raise NotFoundError('release with ID %s not found' % rel_id)
-        if not series.season_in_show(show_id, season_id):
+        if not db.season_in_show(show_id, season_id):
             raise BadRequest('season with id %s does not belong to show %s' % (season_id, show_id))
-        if not series.release_in_season(season_id, rel_id):
+        if not db.release_in_season(season_id, rel_id):
             raise BadRequest('release id %s does not belong to episode %s' % (rel_id, season_id))
 
         if not release.downloaded:
@@ -882,36 +985,37 @@ class SeriesSeasonReleaseAPI(APIResource):
         release.downloaded = False
 
         rsp = jsonify(release.to_dict())
-        rsp.headers.extend({
-            'Series-ID': show_id,
-            'Season-ID': season_id
-        })
+        rsp.headers.extend({'Series-ID': show_id, 'Season-ID': season_id})
         return rsp
 
 
 @api.response(NotFoundError)
 @api.response(BadRequest)
 @series_api.route('/<int:show_id>/episodes/<int:ep_id>/releases/')
-@api.doc(params={'show_id': 'ID of the show', 'ep_id': 'Episode ID'},
-         description='Releases are any seen entries that match the episode. \n'
-                     'The \'Series-ID\' header will be appended to the result headers.\n'
-                     'The \'Episode-ID\' header will be appended to the result headers.')
+@api.doc(
+    params={'show_id': 'ID of the show', 'ep_id': 'Episode ID'},
+    description='Releases are any seen entries that match the episode. \n'
+    'The \'Series-ID\' header will be appended to the result headers.\n'
+    'The \'Episode-ID\' header will be appended to the result headers.',
+)
 class SeriesEpisodeReleasesAPI(APIResource):
     @etag
     @api.response(200, 'Releases retrieved successfully for episode', episode_release_list_schema)
-    @api.doc(description='Get all matching releases for a specific episode of a specific show.',
-             parser=release_list_parser)
+    @api.doc(
+        description='Get all matching releases for a specific episode of a specific show.',
+        parser=release_list_parser,
+    )
     def get(self, show_id, ep_id, session):
         """ Get all episodes releases by show ID and episode ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            episode = series.episode_by_id(ep_id, session)
+            episode = db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
 
         args = release_list_parser.parse_args()
@@ -940,14 +1044,14 @@ class SeriesEpisodeReleasesAPI(APIResource):
             'sort_by': sort_by,
             'descending': descending,
             'episode': episode,
-            'session': session
+            'session': session,
         }
 
         # Total number of releases
-        total_items = series.get_episode_releases(count=True, **kwargs)
+        total_items = db.get_episode_releases(count=True, **kwargs)
 
         # Release items
-        release_items = [release.to_dict() for release in series.get_episode_releases(**kwargs)]
+        release_items = [release.to_dict() for release in db.get_episode_releases(**kwargs)]
 
         # Total number of pages
         total_pages = int(ceil(total_items / float(per_page)))
@@ -973,48 +1077,62 @@ class SeriesEpisodeReleasesAPI(APIResource):
         return rsp
 
     @api.response(200, 'Successfully deleted all releases for episode', model=base_message_schema)
-    @api.doc(description='Delete all releases for a specific episode of a specific show.',
-             parser=release_delete_parser)
+    @api.doc(
+        description='Delete all releases for a specific episode of a specific show.',
+        parser=release_delete_parser,
+    )
     def delete(self, show_id, ep_id, session):
         """ Deletes all episodes releases by show ID and episode ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            episode = series.episode_by_id(ep_id, session)
+            episode = db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
 
         args = release_delete_parser.parse_args()
         downloaded = args.get('downloaded') is True if args.get('downloaded') is not None else None
         release_items = []
         for release in episode.releases:
-            if downloaded and release.downloaded or downloaded is False and not release.downloaded or not downloaded:
+            if (
+                downloaded
+                and release.downloaded
+                or downloaded is False
+                and not release.downloaded
+                or not downloaded
+            ):
                 release_items.append(release)
 
         for release in release_items:
             if args.get('forget'):
                 fire_event('forget', release.title)
-            series.delete_episode_release_by_id(release.id)
-        return success_response('successfully deleted all releases for episode %s from show %s' % (ep_id, show_id))
+            db.delete_episode_release_by_id(release.id)
+        return success_response(
+            'successfully deleted all releases for episode %s from show %s' % (ep_id, show_id)
+        )
 
-    @api.response(200, 'Successfully reset all downloaded releases for episode', model=base_message_schema)
-    @api.doc(description='Resets all of the downloaded releases of an episode, clearing the quality to be downloaded '
-                         'again,')
+    @api.response(
+        200, 'Successfully reset all downloaded releases for episode', model=base_message_schema
+    )
+    @api.doc(
+        description='Resets all of the downloaded releases of an episode, clearing the quality to be downloaded '
+        'again,'
+    )
     def put(self, show_id, ep_id, session):
         """ Marks all downloaded releases as not downloaded """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            episode = series.episode_by_id(ep_id, session)
+            episode = db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
 
         for release in episode.releases:
@@ -1022,94 +1140,104 @@ class SeriesEpisodeReleasesAPI(APIResource):
                 release.downloaded = False
 
         return success_response(
-            'successfully reset download status for all releases for episode %s from show %s' % (ep_id, show_id))
+            'successfully reset download status for all releases for episode %s from show %s'
+            % (ep_id, show_id)
+        )
 
 
 @api.response(NotFoundError)
 @api.response(BadRequest)
 @series_api.route('/<int:show_id>/episodes/<int:ep_id>/releases/<int:rel_id>/')
-@api.doc(params={'show_id': 'ID of the show', 'ep_id': 'Episode ID', 'rel_id': 'Release ID'},
-         description='Releases are any seen entries that match the episode. \n'
-                     'The \'Series-ID\' header will be appended to the result headers.\n'
-                     'The \'Episode-ID\' header will be appended to the result headers.'
-         )
+@api.doc(
+    params={'show_id': 'ID of the show', 'ep_id': 'Episode ID', 'rel_id': 'Release ID'},
+    description='Releases are any seen entries that match the episode. \n'
+    'The \'Series-ID\' header will be appended to the result headers.\n'
+    'The \'Episode-ID\' header will be appended to the result headers.',
+)
 class SeriesEpisodeReleaseAPI(APIResource):
     @etag
     @api.response(200, 'Release retrieved successfully for episode', episode_release_schema)
-    @api.doc(description='Get a specific downloaded release for a specific episode of a specific show')
+    @api.doc(
+        description='Get a specific downloaded release for a specific episode of a specific show'
+    )
     def get(self, show_id, ep_id, rel_id, session):
         """ Get episode release by show ID, episode ID and release ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            series.episode_by_id(ep_id, session)
+            db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
         try:
-            release = series.episode_release_by_id(rel_id, session)
+            release = db.episode_release_by_id(rel_id, session)
         except NoResultFound:
             raise NotFoundError('release with ID %s not found' % rel_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
-        if not series.release_in_episode(ep_id, rel_id):
+        if not db.release_in_episode(ep_id, rel_id):
             raise BadRequest('release id %s does not belong to episode %s' % (rel_id, ep_id))
 
         rsp = jsonify(release.to_dict())
-        rsp.headers.extend({
-            'Series-ID': show_id,
-            'Episode-ID': ep_id
-        })
+        rsp.headers.extend({'Series-ID': show_id, 'Episode-ID': ep_id})
         return rsp
 
     @api.response(200, 'Release successfully deleted', model=base_message_schema)
-    @api.doc(description='Delete a specific releases for a specific episode of a specific show.',
-             parser=delete_parser)
+    @api.doc(
+        description='Delete a specific releases for a specific episode of a specific show.',
+        parser=delete_parser,
+    )
     def delete(self, show_id, ep_id, rel_id, session):
         """ Delete episode release by show ID, episode ID and release ID """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            series.episode_by_id(ep_id, session)
+            db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
         try:
-            release = series.episode_release_by_id(rel_id, session)
+            release = db.episode_release_by_id(rel_id, session)
         except NoResultFound:
             raise NotFoundError('release with ID %s not found' % rel_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
-        if not series.release_in_episode(ep_id, rel_id):
+        if not db.release_in_episode(ep_id, rel_id):
             raise BadRequest('release id %s does not belong to episode %s' % (rel_id, ep_id))
         args = delete_parser.parse_args()
         if args.get('forget'):
             fire_event('forget', release.title)
 
-        series.delete_episode_release_by_id(rel_id)
-        return success_response('successfully deleted release %d from episode %d' % (rel_id, ep_id))
+        db.delete_episode_release_by_id(rel_id)
+        return success_response(
+            'successfully deleted release %d from episode %d' % (rel_id, ep_id)
+        )
 
-    @api.response(200, 'Successfully reset downloaded release status', model=episode_release_schema)
-    @api.doc(description='Resets the downloaded release status, clearing the quality to be downloaded again')
+    @api.response(
+        200, 'Successfully reset downloaded release status', model=episode_release_schema
+    )
+    @api.doc(
+        description='Resets the downloaded release status, clearing the quality to be downloaded again'
+    )
     def put(self, show_id, ep_id, rel_id, session):
         """ Resets a downloaded release status """
         try:
-            series.show_by_id(show_id, session=session)
+            db.show_by_id(show_id, session=session)
         except NoResultFound:
             raise NotFoundError('show with ID %s not found' % show_id)
         try:
-            series.episode_by_id(ep_id, session)
+            db.episode_by_id(ep_id, session)
         except NoResultFound:
             raise NotFoundError('episode with ID %s not found' % ep_id)
         try:
-            release = series.episode_release_by_id(rel_id, session)
+            release = db.episode_release_by_id(rel_id, session)
         except NoResultFound:
             raise NotFoundError('release with ID %s not found' % rel_id)
-        if not series.episode_in_show(show_id, ep_id):
+        if not db.episode_in_show(show_id, ep_id):
             raise BadRequest('episode with id %s does not belong to show %s' % (ep_id, show_id))
-        if not series.release_in_episode(ep_id, rel_id):
+        if not db.release_in_episode(ep_id, rel_id):
             raise BadRequest('release id %s does not belong to episode %s' % (rel_id, ep_id))
 
         if not release.downloaded:
@@ -1117,8 +1245,5 @@ class SeriesEpisodeReleaseAPI(APIResource):
         release.downloaded = False
 
         rsp = jsonify(release.to_dict())
-        rsp.headers.extend({
-            'Series-ID': show_id,
-            'Episode-ID': ep_id
-        })
+        rsp.headers.extend({'Series-ID': show_id, 'Episode-ID': ep_id})
         return rsp
