+++ flexget/components/series/cli.py
@@ -7,19 +7,13 @@ from datetime import timedelta
 
 from colorclass.toggles import disable_all_colors
 
+import flexget.components.series.utils
 from flexget import options
-from flexget import plugin
 from flexget.event import event
 from flexget.manager import Session
 from flexget.terminal import TerminalTable, TerminalTableError, table_parser, colorize, console
 
-try:
-    # NOTE: Importing other plugins is discouraged!
-    from flexget.plugins.filter import series as plugin_series
-except ImportError:
-    raise plugin.DependencyError(
-        issued_by=__name__, missing='series',
-    )
+from . import db
 
 # Environment variables to set defaults for `series list` and `series show`
 ENV_SHOW_SORTBY_FIELD = 'FLEXGET_SERIES_SHOW_SORTBY_FIELD'
@@ -64,25 +58,28 @@ def display_summary(options):
     """
     porcelain = options.table_type == 'porcelain'
     configured = options.configured or os.environ.get(ENV_LIST_CONFIGURED, 'configured')
-    premieres = True if (os.environ.get(ENV_LIST_PREMIERES) == 'yes' or
-                         options.premieres) else False
+    premieres = (
+        True if (os.environ.get(ENV_LIST_PREMIERES) == 'yes' or options.premieres) else False
+    )
     sort_by = options.sort_by or os.environ.get(ENV_LIST_SORTBY_FIELD, 'name')
     if options.order is not None:
         descending = True if options.order == 'desc' else False
     else:
         descending = True if os.environ.get(ENV_LIST_SORTBY_ORDER) == 'desc' else False
     with Session() as session:
-        kwargs = {'configured': configured,
-                  'premieres': premieres,
-                  'session': session,
-                  'sort_by': sort_by,
-                  'descending': descending}
+        kwargs = {
+            'configured': configured,
+            'premieres': premieres,
+            'session': session,
+            'sort_by': sort_by,
+            'descending': descending,
+        }
         if sort_by == 'name':
             kwargs['sort_by'] = 'show_name'
         else:
             kwargs['sort_by'] = 'last_download_date'
 
-        query = plugin_series.get_series_summary(**kwargs)
+        query = db.get_series_summary(**kwargs)
         header = ['Name', 'Begin', 'Last Encountered', 'Age', 'Downloaded', 'Identified By']
         for index, value in enumerate(header):
             if value.lower() == options.sort_by:
@@ -97,12 +94,12 @@ def display_summary(options):
             latest_release = '-'
             age_col = '-'
             episode_id = '-'
-            latest = plugin_series.get_latest_release(series)
+            latest = db.get_latest_release(series)
             identifier_type = series.identified_by
             if identifier_type == 'auto':
                 identifier_type = colorize('yellow', 'auto')
             if latest:
-                behind = plugin_series.new_entities_after(latest)
+                behind = db.new_entities_after(latest)
                 latest_release = get_latest_status(latest)
                 # colorize age
                 age_col = latest.age
@@ -116,11 +113,17 @@ def display_summary(options):
                 episode_id = latest.identifier
             if not porcelain:
                 if behind[0] > 0:
-                    name_column += colorize(BEHIND_EP_COLOR, ' {} {} behind'.format(behind[0], behind[1]))
+                    name_column += colorize(
+                        BEHIND_EP_COLOR, ' {} {} behind'.format(behind[0], behind[1])
+                    )
 
-            table_data.append([name_column, begin, episode_id, age_col, latest_release, identifier_type])
+            table_data.append(
+                [name_column, begin, episode_id, age_col, latest_release, identifier_type]
+            )
     try:
-        table = TerminalTable(options.table_type, table_data, wrap_columns=[3], drop_columns=[4, 3, 2])
+        table = TerminalTable(
+            options.table_type, table_data, wrap_columns=[3], drop_columns=[4, 3, 2]
+        )
         console(table.output)
     except TerminalTableError as e:
         console('ERROR: %s' % str(e))
@@ -135,9 +138,9 @@ def display_summary(options):
 def begin(manager, options):
     series_name = options.series_name
     series_name = series_name.replace(r'\!', '!')
-    normalized_name = plugin_series.normalize_series_name(series_name)
+    normalized_name = flexget.components.series.utils.normalize_series_name(series_name)
     with Session() as session:
-        series = plugin_series.shows_by_exact_name(normalized_name, session)
+        series = db.shows_by_exact_name(normalized_name, session)
         if options.forget:
             if not series:
                 console('Series `%s` was not found in the database.' % series_name)
@@ -151,20 +154,22 @@ def begin(manager, options):
             ep_id = options.episode_id
             if not series:
                 console('Series not yet in database. Adding `%s`.' % series_name)
-                series = plugin_series.Series()
+                series = db.Series()
                 series.name = series_name
                 session.add(series)
             else:
                 series = series[0]
             try:
-                _, entity_type = plugin_series.set_series_begin(series, ep_id)
+                _, entity_type = db.set_series_begin(series, ep_id)
             except ValueError as e:
                 console(e)
             else:
                 if entity_type == 'season':
                     console('`%s` was identified as a season.' % ep_id)
                     ep_id += 'E01'
-                console('Releases for `%s` will be accepted starting with `%s`.' % (series.name, ep_id))
+                console(
+                    'Releases for `%s` will be accepted starting with `%s`.' % (series.name, ep_id)
+                )
                 session.commit()
             manager.config_changed()
 
@@ -174,15 +179,18 @@ def remove(manager, options, forget=False):
     if options.episode_id:
         for identifier in options.episode_id:
             try:
-                plugin_series.remove_series_entity(name, identifier, forget)
+                db.remove_series_entity(name, identifier, forget)
             except ValueError as e:
                 console(e.args[0])
             else:
-                console('Removed entities(s) matching `%s` from series `%s`.' % (identifier, name.capitalize()))
+                console(
+                    'Removed entities(s) matching `%s` from series `%s`.'
+                    % (identifier, name.capitalize())
+                )
     else:
         # remove whole series
         try:
-            plugin_series.remove_series(name, forget)
+            db.remove_series(name, forget)
         except ValueError as e:
             console(e.args[0])
         else:
@@ -218,9 +226,9 @@ def display_details(options):
     else:
         reverse = True if os.environ.get(ENV_SHOW_SORTBY_ORDER) == 'desc' else False
     with Session() as session:
-        name = plugin_series.normalize_series_name(name)
+        name = flexget.components.series.utils.normalize_series_name(name)
         # Sort by length of name, so that partial matches always show shortest matching title
-        matches = plugin_series.shows_by_name(name, session=session)
+        matches = db.shows_by_name(name, session=session)
         if not matches:
             console(colorize(ERROR_COLOR, 'ERROR: Unknown series `%s`' % name))
             return
@@ -228,15 +236,16 @@ def display_details(options):
         series = matches[0]
         table_title = colorize('white', series.name)
         if len(matches) > 1:
-            warning = (colorize('red', ' WARNING: ') +
-                       'Multiple series match to `{}`.\n '
-                       'Be more specific to see the results of other matches:\n\n'
-                       ' {}'.format(name, ', '.join(s.name for s in matches[1:])))
+            warning = (
+                colorize('red', ' WARNING: ') + 'Multiple series match to `{}`.\n '
+                'Be more specific to see the results of other matches:\n\n'
+                ' {}'.format(name, ', '.join(s.name for s in matches[1:]))
+            )
             if not options.table_type == 'porcelain':
                 console(warning)
         header = ['Identifier', 'Last seen', 'Release titles', 'Release Quality', 'Proper']
         table_data = [header]
-        entities = plugin_series.get_all_entities(series, session=session, sort_by=sort_by, reverse=reverse)
+        entities = db.get_all_entities(series, session=session, sort_by=sort_by, reverse=reverse)
         for entity in entities:
             if not entity.releases:
                 continue
@@ -269,12 +278,17 @@ def display_details(options):
             table_data.append(entity_data)
         footer = ' %s \n' % (colorize(DOWNLOADED_RELEASE_COLOR, '* Downloaded'))
         if not series.identified_by:
-            footer += ('\n Series plugin is still learning which episode numbering mode is \n'
-                       ' correct for this series (identified_by: auto).\n'
-                       ' Few duplicate downloads can happen with different numbering schemes\n'
-                       ' during this time.')
+            footer += (
+                '\n Series plugin is still learning which episode numbering mode is \n'
+                ' correct for this series (identified_by: auto).\n'
+                ' Few duplicate downloads can happen with different numbering schemes\n'
+                ' during this time.'
+            )
         else:
-            footer += '\n `%s` uses `%s` mode to identify episode numbering.' % (series.name, series.identified_by)
+            footer += '\n `%s` uses `%s` mode to identify episode numbering.' % (
+                series.name,
+                series.identified_by,
+            )
         begin_text = 'option'
         if series.begin:
             footer += ' \n Begin for `%s` is set to `%s`.' % (series.name, series.begin.identifier)
@@ -295,19 +309,19 @@ def add(manager, options):
     entity_ids = options.entity_id
     quality = options.quality or os.environ.get(ENV_ADD_QUALITY, None)
     series_name = series_name.replace(r'\!', '!')
-    normalized_name = plugin_series.normalize_series_name(series_name)
+    normalized_name = flexget.components.series.utils.normalize_series_name(series_name)
     with Session() as session:
-        series = plugin_series.shows_by_exact_name(normalized_name, session)
+        series = db.shows_by_exact_name(normalized_name, session)
         if not series:
             console('Series not yet in database, adding `%s`' % series_name)
-            series = plugin_series.Series()
+            series = db.Series()
             series.name = series_name
             session.add(series)
         else:
             series = series[0]
         for ent_id in entity_ids:
             try:
-                plugin_series.add_series_entity(session, series, ent_id, quality=quality)
+                db.add_series_entity(session, series, ent_id, quality=quality)
             except ValueError as e:
                 console(e.args[0])
             else:
@@ -319,56 +333,125 @@ def add(manager, options):
 @event('options.register')
 def register_parser_arguments():
     # Register the command
-    parser = options.register_command('series', do_cli, help='View and manipulate the series plugin database')
+    parser = options.register_command(
+        'series', do_cli, help='View and manipulate the series plugin database'
+    )
 
     # Parent parser for subcommands that need a series name
     series_parser = argparse.ArgumentParser(add_help=False)
-    series_parser.add_argument('series_name', help='The name of the series', metavar='<series name>')
+    series_parser.add_argument(
+        'series_name', help='The name of the series', metavar='<series name>'
+    )
 
     # Set up our subparsers
     subparsers = parser.add_subparsers(title='actions', metavar='<action>', dest='series_action')
-    list_parser = subparsers.add_parser('list', parents=[table_parser],
-                                        help='List a summary of the different series being tracked')
-    list_parser.add_argument('configured', nargs='?', choices=['configured', 'unconfigured', 'all'],
-                             help='Limit list to series that are currently in the config or not (default: %(default)s)')
-    list_parser.add_argument('--premieres', action='store_true',
-                             help='limit list to series which only have episode 1 (and maybe also 2) downloaded')
-    list_parser.add_argument('--sort-by', choices=('name', 'age'), help='Choose list sort attribute')
+    list_parser = subparsers.add_parser(
+        'list', parents=[table_parser], help='List a summary of the different series being tracked'
+    )
+    list_parser.add_argument(
+        'configured',
+        nargs='?',
+        choices=['configured', 'unconfigured', 'all'],
+        help='Limit list to series that are currently in the config or not (default: %(default)s)',
+    )
+    list_parser.add_argument(
+        '--premieres',
+        action='store_true',
+        help='limit list to series which only have episode 1 (and maybe also 2) downloaded',
+    )
+    list_parser.add_argument(
+        '--sort-by', choices=('name', 'age'), help='Choose list sort attribute'
+    )
     order = list_parser.add_mutually_exclusive_group(required=False)
-    order.add_argument('--descending', dest='order', action='store_const', const='desc',
-                       help='Sort in descending order')
-    order.add_argument('--ascending', dest='order', action='store_const', const='asc', help='Sort in ascending order')
+    order.add_argument(
+        '--descending',
+        dest='order',
+        action='store_const',
+        const='desc',
+        help='Sort in descending order',
+    )
+    order.add_argument(
+        '--ascending',
+        dest='order',
+        action='store_const',
+        const='asc',
+        help='Sort in ascending order',
+    )
 
-    show_parser = subparsers.add_parser('show', parents=[series_parser, table_parser],
-                                        help='Show the releases FlexGet has seen for a given series')
-    show_parser.add_argument('--sort-by', choices=('age', 'identifier'),
-                             help='Choose releases list sort: age (default) or identifier')
+    show_parser = subparsers.add_parser(
+        'show',
+        parents=[series_parser, table_parser],
+        help='Show the releases FlexGet has seen for a given series',
+    )
+    show_parser.add_argument(
+        '--sort-by',
+        choices=('age', 'identifier'),
+        help='Choose releases list sort: age (default) or identifier',
+    )
     show_order = show_parser.add_mutually_exclusive_group(required=False)
-    show_order.add_argument('--descending', dest='order', action='store_const', const='desc',
-                            help='Sort in descending order')
-    show_order.add_argument('--ascending', dest='order', action='store_const', const='asc',
-                            help='Sort in ascending order')
+    show_order.add_argument(
+        '--descending',
+        dest='order',
+        action='store_const',
+        const='desc',
+        help='Sort in descending order',
+    )
+    show_order.add_argument(
+        '--ascending',
+        dest='order',
+        action='store_const',
+        const='asc',
+        help='Sort in ascending order',
+    )
 
-    begin_parser = subparsers.add_parser('begin', parents=[series_parser],
-                                         help='set the episode to start getting a series from')
+    begin_parser = subparsers.add_parser(
+        'begin', parents=[series_parser], help='set the episode to start getting a series from'
+    )
     begin_opts = begin_parser.add_mutually_exclusive_group(required=True)
-    begin_opts.add_argument('--forget', dest='forget', action='store_true', help='Forget begin episode', required=False)
-    begin_opts.add_argument('episode_id', metavar='<episode ID>',
-                            help='Episode ID to start getting the series from (e.g. S02E01, 2013-12-11, or 9, '
-                                 'depending on how the series is numbered). You can also enter a season ID such as '
-                                 ' S02.', nargs='?', default='')
+    begin_opts.add_argument(
+        '--forget', dest='forget', action='store_true', help='Forget begin episode', required=False
+    )
+    begin_opts.add_argument(
+        'episode_id',
+        metavar='<episode ID>',
+        help='Episode ID to start getting the series from (e.g. S02E01, 2013-12-11, or 9, '
+        'depending on how the series is numbered). You can also enter a season ID such as '
+        ' S02.',
+        nargs='?',
+        default='',
+    )
 
-    addshow_parser = subparsers.add_parser('add', parents=[series_parser],
-                                           help='Add episode(s) and season(s) to series history')
-    addshow_parser.add_argument('entity_id', nargs='+', default=None, metavar='<entity_id>',
-                                help='Episode or season entity ID(s) to add')
-    addshow_parser.add_argument('--quality', default=None, metavar='<quality>',
-                                help='Quality string to be stored for all entity ID(s)')
+    addshow_parser = subparsers.add_parser(
+        'add', parents=[series_parser], help='Add episode(s) and season(s) to series history'
+    )
+    addshow_parser.add_argument(
+        'entity_id',
+        nargs='+',
+        default=None,
+        metavar='<entity_id>',
+        help='Episode or season entity ID(s) to add',
+    )
+    addshow_parser.add_argument(
+        '--quality',
+        default=None,
+        metavar='<quality>',
+        help='Quality string to be stored for all entity ID(s)',
+    )
 
-    forget_parser = subparsers.add_parser('forget', parents=[series_parser],
-                                          help='Removes episodes or whole series from the entire database '
-                                               '(including seen plugin)')
-    forget_parser.add_argument('episode_id', nargs='*', default=None, help='Entity ID(s) to forget (optional)')
-    delete_parser = subparsers.add_parser('remove', parents=[series_parser],
-                                          help='Removes episodes or whole series from the series database only')
-    delete_parser.add_argument('episode_id', nargs='*', default=None, help='Episode ID to forget (optional)')
+    forget_parser = subparsers.add_parser(
+        'forget',
+        parents=[series_parser],
+        help='Removes episodes or whole series from the entire database '
+        '(including seen plugin)',
+    )
+    forget_parser.add_argument(
+        'episode_id', nargs='*', default=None, help='Entity ID(s) to forget (optional)'
+    )
+    delete_parser = subparsers.add_parser(
+        'remove',
+        parents=[series_parser],
+        help='Removes episodes or whole series from the series database only',
+    )
+    delete_parser.add_argument(
+        'episode_id', nargs='*', default=None, help='Episode ID to forget (optional)'
+    )
