+++ flexget/components/series/series_begin.py
@@ -6,13 +6,7 @@ from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from flexget import plugin
 from flexget.event import event
 
-try:
-    # NOTE: Importing other plugins is discouraged!
-    from flexget.plugins.filter import series as plugin_series
-except ImportError:
-    raise plugin.DependencyError(
-        issued_by=__name__, missing='series',
-    )
+from . import db
 
 log = logging.getLogger('set_series_begin')
 
@@ -34,17 +28,24 @@ class SetSeriesBegin(object):
             return
         for entry in task.accepted:
             if entry.get('series_name') and entry.get('series_id'):
-                fshow = task.session.query(plugin_series.Series).filter(
-                    plugin_series.Series.name == entry['series_name']).first()
+                fshow = (
+                    task.session.query(db.Series)
+                    .filter(db.Series.name == entry['series_name'])
+                    .first()
+                )
                 if not fshow:
-                    fshow = plugin_series.Series()
+                    fshow = db.Series()
                     fshow.name = entry['series_name']
                     task.session.add(fshow)
                 try:
-                    plugin_series.set_series_begin(fshow, entry['series_id'])
+                    db.set_series_begin(fshow, entry['series_id'])
                 except ValueError as e:
-                    log.error('An error occurred trying to set begin for %s: %s', entry['series_name'], e)
-                log.info('First episode for "%s" set to %s', entry['series_name'], entry['series_id'])
+                    log.error(
+                        'An error occurred trying to set begin for %s: %s', entry['series_name'], e
+                    )
+                log.info(
+                    'First episode for "%s" set to %s', entry['series_name'], entry['series_id']
+                )
 
 
 @event('plugin.register')
