+++ flexget/components/series/series_forget.py
@@ -6,13 +6,7 @@ from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from flexget import plugin
 from flexget.event import event
 
-try:
-    # NOTE: Importing other plugins is discouraged!
-    from flexget.plugins.filter import series as plugin_series
-except ImportError:
-    raise plugin.DependencyError(
-        issued_by=__name__, missing='series',
-    )
+from . import db
 
 log = logging.getLogger('series_forget')
 
@@ -26,11 +20,16 @@ class OutputSeriesRemove(object):
         for entry in task.accepted:
             if 'series_name' in entry and 'series_id' in entry:
                 try:
-                    plugin_series.remove_series_entity(entry['series_name'], entry['series_id'])
-                    log.info('Removed episode `%s` from series `%s` download history.' %
-                             (entry['series_id'], entry['series_name']))
+                    db.remove_series_entity(entry['series_name'], entry['series_id'])
+                    log.info(
+                        'Removed episode `%s` from series `%s` download history.'
+                        % (entry['series_id'], entry['series_name'])
+                    )
                 except ValueError:
-                    log.debug('Series (%s) or id (%s) unknown.' % (entry['series_name'], entry['series_id']))
+                    log.debug(
+                        'Series (%s) or id (%s) unknown.'
+                        % (entry['series_name'], entry['series_id'])
+                    )
 
 
 @event('plugin.register')
